Add the ksql-examples docker image
This patch enables building the ksql-examples docker image. This is a
bit tricky because we want to build an image named ksql-examples but
have an artifact id of cp-ksql-examples, to avoid collidign with the
ksql-examples artifact from the ksql repo. So we have to update the
maven dockerfile plug-in and Dockerfile accordingly.